refactor: use netlify/functions
