--- src/utils/exists.js
@@ -1,7 +1,0 @@
-import fs from 'fs';
-
-const exists = (file) => {
-  return fs.existsSync(file);
-};
-
-export default exists;
